Released 2.1.2
git-svn-id: https://plugins.svn.wordpress.org/nextgen-gallery/trunk@1202400 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ products/photocrati_nextgen/modules/fs/package.module.fs.php
@@ -222,6 +222,9 @@ class Mixin_Fs_Instance_Methods extends Mixin
                 @unlink($abspath);
             } else {
                 foreach (scandir($abspath) as $relpath) {
+                    if (in_array($relpath, array('.', '..'))) {
+                        continue;
+                    }
                     $sub_abspath = $this->join_paths($abspath, $relpath);
                     $this->object->delete($sub_abspath);
                 }
